ifsc_django/requirements.txt
anyio==3.6.1
apparmor==3.0.4
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.5
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
Babel==2.10.1
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==4.1.0
btrfsutil==5.17
CacheControl==0.12.6
cffi==1.15.0
chardet==4.0.0
click==8.1.3
colorama==0.4.4
contextlib2==0.6.0.post1
cryptography==37.0.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
deprecation==2.1.0
distlib==0.3.4
distro==1.7.0
Django==4.0.6
django-excel==0.0.10
django-filter==22.1
django-redis==5.2.0
djangorestframework==3.13.1
dnspython==2.2.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.6.0
Flask==2.1.3
Flask-Excel==0.0.7
GDAL==3.4.3
html5lib==1.1
idna==3.3
importlib-metadata==4.8.1
ipykernel==6.13.0
ipython==8.3.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.0.3
json5==0.9.8
jsonschema==3.2.0
jupyter-client==7.3.1
jupyter-core==4.10.0
jupyter-packaging==0.12.0
jupyter-server==1.17.0
jupyterlab==3.4.2
jupyterlab-pygments==0.2.2
jupyterlab-server==2.14.0
lensfun==0.3.3
LibAppArmor==3.0.4
lml==0.1.0
louis==3.21.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
meld==3.20.4
mistune==0.8.4
more-itertools==8.12.0
msgpack==1.0.3
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.4
notebook==6.4.11
notebook-shim==0.1.0
numpy==1.22.3
openpyxl==3.0.10
ordered-set==4.0.2
OWSLib==0.25.0
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.2
pep517==0.12.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.2
ply==3.11
progress==1.6
prometheus-client==0.14.1
prompt-toolkit==3.0.28
psutil==5.9.0
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pwquality==1.4.4
pycairo==1.21.0
pycparser==2.21
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
pyexcel-xls==0.7.0
pyexcel-xlsx==0.6.0
Pygments==2.12.0
PyGObject==3.42.1
PyICU==2.9
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyproj==3.0.1
PyQt5==5.15.6
PyQt5-sip==12.10.1
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
pyzmq==22.3.0
QScintilla==2.13.3
ranger-fm==1.9.3
redis==4.3.4
requests==2.27.1
requests-unixsocket==0.3.0
resolvelib==0.5.5
retrying==1.3.3
Send2Trash==1.8.0
setproctitle==1.2.3
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
sqlparse==0.4.2
stack-data==0.2.0
TBB==0.2
team==1.0
terminado==0.15.0
testpath==0.6.0
texttable==1.6.4
tinycss2==1.1.1
TLPUI==1.5.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.10.2
tornado==6.1
traitlets==5.2.1.post0
urllib3==1.26.9
virtualenv==20.14.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
Werkzeug==2.2.0
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0
zipp==3.8.0
